Step 7 Complete!
I changed the way the prices work from being constants to only being constants
at the start, but then changing them so I can increase the price as the player
buys more of the same upgrade.

I also made changes to update the UI to reflect what the changes look like,
such as the buttons displaying the increased prices.